refactor: extract request-side Ctx methods into ctx_request.go
